+++ data/duke.txt
@@ -1,1 +1,2 @@
 T | 0 | go home
+D | 0 | finish work | 29-01-2023
