+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,19 @@
+package duke;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList(ArrayList<Task> t) {
+        this.tasks = t;
+    }
+
+    public TaskList() {
+        this.tasks = new ArrayList<Task>();
+    }
+
+    public ArrayList<Task> getTasks() {
+        return this.tasks;
+    }
+}
