+++ src/main/java/duke/Todo.java
@@ -1,0 +1,18 @@
+package duke;
+
+public class Todo extends Task {
+
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toSavedString() {
+        return "T | " + super.toSavedString();
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
